+++ packages/jaspr/CHANGELOG.md
@@ -3,6 +3,7 @@
 - **BREAKING** Migrate all packages to `package:web`, replacing `dart:html`.
 
 - Add `InheritedModel<T>` similar to Flutters [InheritedModel](https://api.flutter.dev/flutter/widgets/InheritedModel-class.html)
+- Improved html formatting on the server to not introduce unwanted whitespaces.
 - Fixed server issue during tests where the web directory would never resolve.
 - Fixed issue with unhandled parameter types of client components.